[Gold III] Title: 우주신과의 교감, Time: 632 ms, Memory: 44388 KB -BaekjoonHub

+++ 백준/Gold/1774. 우주신과의 교감/우주신과의 교감.java
@@ -2,14 +2,14 @@
 //MST - 크루스칼(유니온 파인)로 풀었다.
 //Node class에 평소에 int형을 compareTo로 비교하여 정렬했는데 이번에 double형을 사용하며
 //정렬이 안되는 문제가 발생 - Double.compare() 함수로 해결
-//double 실수형 더하는 과정에서 근사값문제로 BigDecimal을 사용해야 된다는 것을 알게됨
+//double 실수형 더하는 과정에서 근사값문제로 BigDecimal을 사용하는 방향으로 감
 // BigDecimal을 처음 써보게 되는 문제였음 BigDecimal 어렵..
 // 틀린 부분 수정 
 // -1. 거리 계산 부분을 int형에서 long으로 변경
 // -2. 이미 연결되어 있는 좌표 두 개의 번호가 오름차순으로 주어진다는 보장이 없어서 union 수정
+// ###########틀린 부분 수정############ 
 // -3. 주요 수정 사항:
 //BigDecimal 제거: BigDecimal을 제거하고, double을 사용해 계산 후 출력 시에 소수점 2자리까지 출력하도록 변경했습니다.
-//M개의 연결된 통로 처리: 이미 연결된 통로는 union(tmpX, tmpY)로 처리하고 있습니다. 이 부분은 문제없지만, 중복된 연결이 발생하지 않도록 오름차순 처리 보장이 있는지 확인했습니다.
 //소수점 출력 처리: String.format("%.2f")을 사용해 소수점 2자리까지 결과를 정확하게 출력합니다.
 import java.io.*;
 import java.util.*;
@@ -45,7 +45,7 @@ public class Main {
 		for(int i = 1; i <= N; i++) {
 			parent[i] = i;
 		}
-		long[][] arr = new long[N+1][2];
+		int[][] arr = new int[N+1][2];
 		nodeList = new ArrayList<>();
 		
 		//좌표 x, y를 우선 2차원 배열에 저장한다.
@@ -69,7 +69,6 @@ public class Main {
 				double x_dist = Math.pow(arr[to][0] - arr[from][0], 2);
 				double y_dist = Math.pow(arr[to][1] - arr[from][1], 2);
 				double cost = Math.sqrt(x_dist + y_dist);
-//	            double cost = Math.sqrt(Math.pow(arr[to][0] - arr[from][0], 2) + Math.pow(arr[to][1] - arr[from][1], 2));
 //				System.out.println(cost);
 				nodeList.add(new Node(from, to, cost)); //유니온 파인드를 사용하기 위해 Node class 형태로 넣는다.
 			}
